refactor(brand): resolve react doctor findings on brand form
- Stabilize onSavingChange/onSavedAtChange with useCallback so the form's
  unmount-reset effect runs once instead of looping.
- Replace the navigator.languages mount effect with a useSyncExternalStore
  hook to drop the set-state-in-effect error and the hydration flash.
- Lazily initialize the lastSavedData ref instead of stringifying every render.
- Add an accessible label to the website input.
- Move form logic into useBrandForm and split the JSX into section components.
